Fixed date and time for notifications
