+++ Assets/Dev/Scripts/Control Room/NotificationPanel.cs
@@ -4,23 +4,28 @@ using UnityEngine;
 using UnityEngine.UI;
 
 public class NotificationPanel : MonoBehaviour {
-    private Text username;
-    private Text message;
-    private Image mediaPlaform;
     //image needs to be public to be able to clone
     public Sprite image;
     public GameObject panelImage;
+    public GameObject notificationMenu;
+    public GameObject Timestamp;
+    public bool isFavorite;
+    private Text username;
+    private Text message;
+    private Image mediaPlaform;
     private GameObject imageButton;
     private Image favoriteButton;
-    public GameObject notificationMenu;
-    public bool isFavorite;
+    private Text Date;
 
     public void SetGameObjects(){
+        imageButton = this.transform.Find("Show Button").gameObject;
+    }
+    public void SetComponents(){
         username = this.transform.Find("UserName").GetComponent<Text>();
         message = this.transform.Find("Message").GetComponent<Text>();
         mediaPlaform = this.transform.Find("MediaPlatform").GetComponent<Image>();
         favoriteButton = this.transform.Find("Favorite Button").GetComponent<Image>();
-        imageButton = this.transform.Find("Show Button").gameObject;
+        Date = this.transform.Find("Date").GetComponent<Text>();
     }
     
     public void SetName(string name) {
@@ -41,6 +46,11 @@ public class NotificationPanel : MonoBehaviour {
             imageButton.SetActive(true);
 
         }
+    }
+    public void SetTime(){
+        Text date = Timestamp.transform.Find("Date").GetComponent<Text>();
+        Text time = Timestamp.transform.Find("Time").GetComponent<Text>();
+        Date.text = string.Format("{0} \n{1}", time.text, date.text);
     }
     public void ShowImage() {
         panelImage.GetComponent<Image>().sprite = image;
