--- Assets/Dev/Scripts/Control Room/Timer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 11f9cf72645b244f4a5b54ff6a837413
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
